+++ lagrange-core/src/main/java/com/jinsulive/lagrange/core/event/message/GroupMessageEvent.java
@@ -18,7 +18,7 @@ public class GroupMessageEvent {
     private Long time;
     private Long selfId;
     private PostType postType;
-    private Integer messageId;
+    private Long messageId;
     private Long groupId;
     private Long userId;
     private String anonymous;
@@ -95,11 +95,11 @@ public class GroupMessageEvent {
         this.postType = postType;
     }
 
-    public Integer getMessageId() {
+    public Long getMessageId() {
         return messageId;
     }
 
-    public void setMessageId(Integer messageId) {
+    public void setMessageId(Long messageId) {
         this.messageId = messageId;
     }
 
@@ -200,7 +200,7 @@ public class GroupMessageEvent {
         private Long time;
         private Long selfId;
         private PostType postType;
-        private Integer messageId;
+        private Long messageId;
         private Long groupId;
         private Long userId;
         private String anonymous;
@@ -231,7 +231,7 @@ public class GroupMessageEvent {
             return this;
         }
 
-        public Builder messageId(Integer messageId) {
+        public Builder messageId(Long messageId) {
             this.messageId = messageId;
             return this;
         }
